add conditional to migration file

+++ db/migrations/testing/20190705143327_add_salesforce_id_rename_primary.js
@@ -7,9 +7,15 @@ exports.up = function(knex) {
           .then(function(exists) {
             if (!exists) {
               return knex.schema.table("submissions", function(table) {
-                table.uuid("salesforce_id").notNullable().unique;
-                table.renameColumn("submission_id", "id");
-                table.dropColumn("contact_id");
+                return knex.schema
+                  .hasColumn("submissions", "submission_id")
+                  .then(function(exists) {
+                    if (exists) {
+                      table.uuid("salesforce_id").notNullable().unique;
+                      table.renameColumn("submission_id", "id");
+                      table.dropColumn("contact_id");
+                    }
+                  });
               });
             }
           });
